+++ barapost/barapost-binning.py
@@ -1,9 +1,9 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 
-__version__ = "4.7.a"
+__version__ = "4.7.b"
 # Year, month, day
-__last_update_date__ = "2021-01-09"
+__last_update_date__ = "2021-01-11"
 
 # |===== Check python interpreter version =====|